Apply pytest everywhere.
I modified the setup.cfg file so that pytest would discover tests everywhere. Note that, for pytest, files with the prefix "test_" or the postix "_test" are assumed to be files with pytests in them. I renamed files that were named this way but were not pytests to use "check" instead to avoid collisions with pytest.